// src/types/types.ts
export type CurrentSettings = {
  tempUnit: string;
  speedUnit: string;
  dateFormat: string;
  // language: string;
};
// ...
export type Option = { value: string; label: string };

export interface SettingsItemProps {
  label: string;
  id: string;
  options: Option[];
  value: string;
  onChange: (event: React.ChangeEvent<HTMLSelectElement>) => void;
}

export type SettingKey = keyof CurrentSettings;